requirements.txt
python-fasthtml
fa6_icons
datasets
pandas
torch
numpy>=1.19.0,<2.0.0
scikit-learn